requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
sqlalchemy

# development
black